Move SystemAudioDevicesChangeEventTarget to Use Cases
Issue #200
